+++ src/utils/completeOptions.test.ts
@@ -142,6 +142,7 @@ describe('completeOptions', () => {
         useHashInUrlAsPathname: false,
         cookieStrategy: CookieStrategy.default,
         cookieWriter: null,
+        inlineResources: true,
       }
       expect(completed).toStrictEqual(expected)
     })
@@ -171,6 +172,7 @@ describe('completeOptions', () => {
         useHashInUrlAsPathname: false,
         cookieStrategy: CookieStrategy.default,
         cookieWriter: null,
+        inlineResources: true,
       }
       expect(completed).toStrictEqual(expected)
     })
@@ -204,6 +206,7 @@ describe('completeOptions', () => {
           useHashInUrlAsPathname: false,
           cookieStrategy: CookieStrategy.default,
           cookieWriter: null,
+          inlineResources: true,
         }
         expect(completed).toStrictEqual(expected)
       })
@@ -231,6 +234,7 @@ describe('completeOptions', () => {
           useHashInUrlAsPathname: false,
           cookieStrategy: CookieStrategy.default,
           cookieWriter: null,
+          inlineResources: true,
         }
         expect(completed).toStrictEqual(expected)
       })
@@ -258,6 +262,7 @@ describe('completeOptions', () => {
           useHashInUrlAsPathname: false,
           cookieStrategy: CookieStrategy.default,
           cookieWriter: null,
+          inlineResources: true,
         }
         expect(completed).toStrictEqual(expected)
       })
@@ -285,6 +290,7 @@ describe('completeOptions', () => {
           useHashInUrlAsPathname: false,
           cookieStrategy: CookieStrategy.default,
           cookieWriter: null,
+          inlineResources: true,
         }
         expect(completed).toStrictEqual(expected)
       })