+++ CHANGELOG.md
@@ -5,13 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased](https://github.com/Smartesting/gravity-data-collector/compare/v6.0.0...main)
+## [Unreleased](https://github.com/Smartesting/gravity-data-collector/compare/v6.1.0...main)
 
 ### Info
 
 ### Added
-
-- `inlineResources` option to inline (or not) the style of the recorded snapshots
 
 ### Changed
 
@@ -20,6 +18,12 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### Removed
 
 ### Fixed
+
+## [6.1.0](https://github.com/Smartesting/gravity-data-collector/compare/v6.0.0...v6.1.0)
+
+### Added
+
+- `inlineResources` option to inline (or not) the style of the recorded snapshots
 
 ## [6.0.0](https://github.com/Smartesting/gravity-data-collector/compare/v5.1.0...v6.0.0)